Anomaly_inference_flask_app/requirements.txt
#### Patchcore

click>= 8.0.3
cudatoolkit>= 10.2
# faiss-cpu
faiss-gpu-cu12
matplotlib>= 3.5.0
pillow>= 8.4.0
pretrainedmodels>= 0.7.4
torch>= 1.10.0
scikit-image>= 0.18.3
scikit-learn>= 1.0.1
scipy>= 1.7.1
torchvision>= 0.11.1
tqdm>= 4.62.3
timm


### Student-Teacher
opencv-python
# tritonclient[all]
# protobuf==3.20.1


# FLASK APP

Flask
Flask-RESTful
Flask-Cors